Enable off-line analysis.
Instead of always looking at '/proc/interrupts' allow the user to pass
as an argument a file, which could for instance be a sample of
'/proc/interrupts' collected at some point in time.

In fact, all the per-node statistics, the filtering, etc., makes perfect
sense for this use-case too.

Of course, in this case, only batch mode can be used (and is in fact
forced). And iterations do not really makes sense.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
---
Changes from v1:
* fixed style issues