Fix using '--totals' and '--sort i' at the same time.
Sorting with 'i' and specifying '--totals' cause the script to quit,
with this trace:

Traceback (most recent call last):
  File "./irqstat", line 341, in <module>
    sys.exit(main(sys.argv))
  File "./irqstat", line 332, in main
    options.node, options.zero, options.filter, options.file1)
  File "./irqstat", line 190, in display_itop
    rows = sorted(irqs.values(), key=sort_func, reverse=rev)
  File "./irqstat", line 185, in sort_func
    return int(val['num'])
ValueError: invalid literal for int() with base 10: 'LOC'

Fix it. After this commit, when sorting with 'i', the lines enabled by
also using '--totals' (which are not associated with an interrupt
number) always go to the bottom.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>