Implement off-line comparison.
After off-line analysis of a sample of '/proc/interrupts' saved into a
file, what about if I have two samples (in two files) and I want to
compare the numbers in them.

Basically, it's the same that irqstat already does, between two
different iterations, when doing online monitoring. With the only
difference that we again do not really do any real iteration, we just do
all the stats and the differences/comparisons between two samples.

This can be really useful if one has, for instance, a file called
'interrupts_before', which is a sample of '/proc/interrupts' taken
before a test, a benchmark, etc. And 'interrupts_after' taken after the
end of such activity.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
---
Changes from v1:
* fixed style issues
* get rid of temporary comment