Support for off-line NUMA information.
Especially if we are doing off-line analysis, it may happen that we do
it on a different machine from where the samples of '/proc/interrupts'
where collected. In which case, calling 'numactl --hardware' will not
give use the info we need, for building the per-NUMA node IRQ stats.

Enable the tool to deal with off-line input for NUMA information as well
(in the form of a file which has the content of 'numactl --hardware' run
on the appropriate machine).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
---
Changes from v1:
* fixed quite a few style issue